+++ src/main/java/org/swyg/greensumer/dto/request/StoreCreateRequest.java
@@ -14,13 +14,13 @@ public class StoreCreateRequest {
     private String type;
     private String description;
     private String hours;
-    private List<Integer> images;
+    private List<Long> images;
     private String address;
     private String roadname;
     private Double lat;
     private Double lng;
 
-    public static StoreCreateRequest of(String name, String type, String description, String hours, List<Integer> images, String address, String roadname, Double lat, Double lng) {
+    public static StoreCreateRequest of(String name, String type, String description, String hours, List<Long> images, String address, String roadname, Double lat, Double lng) {
         return new StoreCreateRequest(name, type, description, hours, images, address, roadname, lat, lng);
     }
 }
