+++ src/main/java/org/swyg/greensumer/dto/request/UserSignUpRequest.java
@@ -12,7 +12,7 @@ public class UserSignUpRequest {
     private String password;
     private String nickname;
     private String email;
-    private Integer storeId;
+    private Long storeId;
     private String address;
     private Double lat;
     private Double lng;
@@ -21,7 +21,7 @@ public class UserSignUpRequest {
         return new UserSignUpRequest(username, password, email, nickname, null, null, null, null);
     }
 
-    public static UserSignUpRequest of(String username, String password, String email, String nickname, Integer storeId, String address, Double lat, Double lng) {
+    public static UserSignUpRequest of(String username, String password, String email, String nickname, Long storeId, String address, Double lat, Double lng) {
         return new UserSignUpRequest(username, password, email, nickname, storeId, address, lat, lng);
     }
